test(mocks): add mock data for quiz tests and refactor existing mocks
- Introduced a new mockData.ts file containing mock quizzes and user quiz answers for improved test coverage.
- Refactored syncUnansweredQuizIdsMocks.ts to utilize the new mock data, enhancing clarity and maintainability of test setups.
- Updated useRef mock implementation to use the shared mockRef from mockData.ts.
