+++ features/answer-quiz/hooks/__tests__/mocks/syncUnansweredQuizIdsMocks.ts
@@ -5,11 +5,14 @@
  */
 
 import { useGlobalContext } from '@/context/GlobalContext'
-import { Tables } from '@/database.types'
 import { useQuery } from '@tanstack/react-query'
-
-// Mock useRef to control the prevUnansweredRef value
-export const mockRef = { current: [99] } // Initialize with a value that will be different from test data
+import {
+  createLimitedMockQuizzes,
+  createMockQuizzes,
+  createMockUserQuizAnswers,
+  createSingleMockUserQuizAnswer,
+  mockRef,
+} from './mockData'
 
 // Setup mocks
 export const setupMocks = (): void => {
@@ -28,125 +31,21 @@ export const setupMocks = (): void => {
 
   jest.mock('react', () => {
     const originalReact = jest.requireActual('react')
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-return
     return {
       ...originalReact,
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-return
       useRef: jest.fn().mockImplementation(() => mockRef),
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call
       useEffect: jest.fn().mockImplementation((fn) => fn()),
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call
       useMemo: jest.fn().mockImplementation((fn) => fn()),
     }
   })
 }
 
-// Mock data for tests
-export const createMockQuizzes = (): Tables<'quizzes'>[] => [
-  {
-    quiz_id: 1,
-    idol_group_id: 1,
-    prompt: 'Q1',
-    choice1: 'A',
-    choice2: 'B',
-    choice3: 'C',
-    choice4: 'D',
-    correct_choice: 1,
-    explanation: 'Explanation 1',
-    quiz_difficulty_id: 1,
-  },
-  {
-    quiz_id: 2,
-    idol_group_id: 1,
-    prompt: 'Q2',
-    choice1: 'A',
-    choice2: 'B',
-    choice3: 'C',
-    choice4: 'D',
-    correct_choice: 2,
-    explanation: 'Explanation 2',
-    quiz_difficulty_id: 1,
-  },
-  {
-    quiz_id: 3,
-    idol_group_id: 1,
-    prompt: 'Q3',
-    choice1: 'A',
-    choice2: 'B',
-    choice3: 'C',
-    choice4: 'D',
-    correct_choice: 3,
-    explanation: 'Explanation 3',
-    quiz_difficulty_id: 1,
-  },
-  {
-    quiz_id: 4,
-    idol_group_id: 1,
-    prompt: 'Q4',
-    choice1: 'A',
-    choice2: 'B',
-    choice3: 'C',
-    choice4: 'D',
-    correct_choice: 4,
-    explanation: 'Explanation 4',
-    quiz_difficulty_id: 1,
-  },
-]
-
-// Shorter quiz list for the last test case
-export const createLimitedMockQuizzes = (): Tables<'quizzes'>[] => [
-  {
-    quiz_id: 1,
-    idol_group_id: 1,
-    prompt: 'Q1',
-    choice1: 'A',
-    choice2: 'B',
-    choice3: 'C',
-    choice4: 'D',
-    correct_choice: 1,
-    explanation: 'Explanation 1',
-    quiz_difficulty_id: 1,
-  },
-  {
-    quiz_id: 2,
-    idol_group_id: 1,
-    prompt: 'Q2',
-    choice1: 'A',
-    choice2: 'B',
-    choice3: 'C',
-    choice4: 'D',
-    correct_choice: 2,
-    explanation: 'Explanation 2',
-    quiz_difficulty_id: 1,
-  },
-]
-
-export const createMockUserQuizAnswers = (): Tables<'user_quiz_answers'>[] => [
-  {
-    app_user_id: 1,
-    quiz_id: 1,
-    selected_choice: 2,
-    is_correct: true,
-    answered_at: '2023-01-01',
-    user_quiz_answer_id: 1,
-  },
-  {
-    app_user_id: 1,
-    quiz_id: 2,
-    selected_choice: 1,
-    is_correct: false,
-    answered_at: '2023-01-01',
-    user_quiz_answer_id: 2,
-  },
-]
-
-export const createSingleMockUserQuizAnswer =
-  (): Tables<'user_quiz_answers'>[] => [
-    {
-      app_user_id: 1,
-      quiz_id: 1,
-      selected_choice: 2,
-      is_correct: true,
-      answered_at: '2023-01-01',
-      user_quiz_answer_id: 1,
-    },
-  ]
+// Re-export the mockRef for tests to use
+export { mockRef }
 
 // Mock query implementations
 export const setupQueryMocks = (
